@@ -12,14 +12,15 @@ classifiers = [
 packages = [{ include = "lib" }]
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.10"
-pandas = "^1.3.4"
+python = ">=3.11,<3.12"
+pandas = "^2.0.3"
 mypy = "^0.991"
 unidecode = "^1.3.6"
+polars = "^0.19.12"
 
 # for data frames schemas / types validation
-pydantic = "^1.10.8"
-pandera = {extras = ["mypy"], version = "^0.15.1"}
+pydantic = "^2.1.1"
+pandera = {extras = ["mypy"], version = "^0.17.0"}
 
 # to simplify development and deployment of cloud functions
 functions-framework = "^3.0.0"
@@ -41,7 +42,7 @@ bracelogger = "^1.1.0"
 
 # for access to AI models
 openai = "^0.27.6"
-langchain = "^0.0.285"
+langchain = "^0.0.275"
 
 # for hugging face
 huggingface_hub = "^0.15.1"
@@ -55,7 +56,13 @@ websocket-client = "^1.6.1"
 # for alibaba
 dashscope = "^1.6.0"
 
+# yival
+yival = {git = "https://github.com/semio/YiVal.git#master"}
+redis = "^5.0.1"
+
 # unsorted (new deps gets listed here)
+langdetect = "^1.0.9"
+
 
 
 [tool.poetry.dev-dependencies]
