@@ -5,18 +5,19 @@
 # for more info
 # Note that most types are str since spreadsheet columns can be formulas
 
-from datetime import datetime
 from typing import Optional
 
 import pandas as pd
 import pandera as pa
 from pandera.engines.pandas_engine import PydanticModel
-from pydantic import BaseModel, Field, validator
+from pydantic import BaseModel, ConfigDict, Field, field_validator
 
 
 class Question(BaseModel):
+    model_config = ConfigDict(coerce_numbers_to_str=True)
+
     include_in_next_evaluation: Optional[bool] = Field(
-        None, title="Include in next evaluation"
+        None, title="Include in next evaluation", validate_default=True
     )
     question_id: Optional[str] = Field(None, title="Question ID")
     language: Optional[str] = Field(None, title="Language")
@@ -24,7 +25,8 @@ class Question(BaseModel):
         None, title="Published version of question"
     )
 
-    @validator("include_in_next_evaluation", pre=True, always=True)
+    @field_validator("include_in_next_evaluation", mode="before")
+    @classmethod
     def default_if_nan(cls, v):  # noqa: N805
         return False if pd.isna(v) else v
 
@@ -36,6 +38,8 @@ class QuestionsDf(pa.DataFrameModel):
 
 
 class QuestionOption(BaseModel):
+    model_config = ConfigDict(coerce_numbers_to_str=True)
+
     question_option_id: Optional[str] = Field(None, title="Question Option ID")
     question_id: Optional[str] = Field(None, title="Question ID")
     language: Optional[str] = Field(None, title="Language")
@@ -53,6 +57,8 @@ class QuestionOptionsDf(pa.DataFrameModel):
 
 
 class PromptVariation(BaseModel):
+    model_config = ConfigDict(coerce_numbers_to_str=True)
+
     include_in_next_evaluation: Optional[bool] = Field(
         None, title="Include in next evaluation"
     )
@@ -78,6 +84,8 @@ class PromptVariationsDf(pa.DataFrameModel):
 
 
 class GenAiModel(BaseModel):
+    model_config = ConfigDict(coerce_numbers_to_str=True, protected_namespaces=())
+
     model_id: Optional[str] = Field(None, title="Model ID")
     vendor: Optional[str] = Field(None, title="Vendor")
     model_name: Optional[str] = Field(None, title="Model name")
@@ -90,6 +98,8 @@ class GenAiModelsDf(pa.DataFrameModel):
 
 
 class GenAiModelConfig(BaseModel):
+    model_config = ConfigDict(coerce_numbers_to_str=True, protected_namespaces=())
+
     include_in_next_evaluation: Optional[bool] = Field(
         None, title="Include in next evaluation"
     )
@@ -107,12 +117,28 @@ class GenAiModelConfigsDf(pa.DataFrameModel):
         coerce = True
 
 
+class Metric(BaseModel):
+    name: Optional[str] = Field(None, title="Name")
+    description: Optional[str] = Field(None, title="Description")
+    prompt: Optional[str] = Field(None, title="Prompt")
+    choices: Optional[str] = Field(None, title="Choices")
+    choice_scores: Optional[str] = Field(None, title="Choice Scores")
+
+
+class MetricsDf(pa.DataFrameModel):
+    class Config:
+        dtype = PydanticModel(Metric)
+        coerce = True
+
+
 class EvalResult(BaseModel):
+    model_config = ConfigDict(coerce_numbers_to_str=True, protected_namespaces=())
+
     question_id: Optional[str] = Field(None, title="Question ID")
     language: Optional[str] = Field(None, title="Language")
     prompt_variation_id: Optional[str] = Field(None, title="Prompt variation ID")
     model_configuration_id: Optional[str] = Field(None, title="Model Configuration ID")
-    last_evaluation_datetime: Optional[datetime] = Field(None, title="Last Evaluation")
+    last_evaluation_datetime: Optional[str] = Field(None, title="Last Evaluation")
     percent_correct: Optional[float] = Field(None, title="Percent Correct")
     percent_wrong: Optional[float] = Field(None, title="Percent Wrong")
     percent_very_wrong: Optional[float] = Field(None, title="Percent Very Wrong")
@@ -128,6 +154,8 @@ class EvalResultsDf(pa.DataFrameModel):
 
 
 class SessionResult(BaseModel):
+    model_config = ConfigDict(coerce_numbers_to_str=True, protected_namespaces=())
+
     session_id: Optional[str] = Field(None, title="Session ID")
     session_time: Optional[str] = Field(None, title="Session Time")
     prompt_variation_id: Optional[str] = Field(None, title="Prompt Variation ID")
