@@ -5,7 +5,7 @@ from typing import Any, Dict, List, Mapping, Optional
 
 from langchain.callbacks.manager import CallbackManagerForLLMRun
 from langchain.llms.base import LLM
-from pydantic import root_validator
+from langchain.pydantic_v1 import root_validator
 from tenacity import (
     retry,
     retry_if_exception_type,
